+++ packages/botonic-plugin-ai-agents/src/types.ts
@@ -1,25 +1,28 @@
 import {
   Agent,
   AgentInputItem,
-  RunContext,
+  RunContext as OpenAIRunContext,
   Tool as OpenAITool,
 } from '@openai/agents'
 import { ZodSchema } from 'zod'
 
 import { OutputMessage, OutputSchema } from './structured-output'
+import { ExitMessage } from './structured-output/exit'
 
 export interface Context {
   authToken: string
 }
 
+export type RunContext = OpenAIRunContext<Context>
+
 export interface CustomTool {
   name: string
   description: string
   schema: ZodSchema
-  func: (input?: any, runContext?: RunContext<Context>) => Promise<any>
+  func: (input?: any, runContext?: RunContext) => Promise<any>
 }
 
-export interface GuardrailFlag {
+export interface GuardrailRule {
   name: string
   description: string
 }
@@ -40,4 +43,14 @@ export interface AiAgentArgs {
 }
 
 export type AgenticInputMessage = AgentInputItem
-export type AgenticOutputMessage = OutputMessage
+export type AgenticOutputMessage = Exclude<OutputMessage, ExitMessage>
+
+export interface RunResult {
+  messages: AgenticOutputMessage[]
+  toolsExecuted: string[]
+  exit: boolean
+  inputGuardrailTriggered: boolean
+  outputGuardrailTriggered: boolean
+}
+
+export type InferenceResponse = RunResult | undefined