+++ packages/botonic-react/src/util/strings.ts
@@ -1,4 +1,8 @@
-export const truncateText = (text, maxLength, ellipsis = '...') => {
+export const truncateText = (
+  text: string,
+  maxLength: number,
+  ellipsis = '...'
+) => {
   if (text.length > maxLength) {
     return text.substring(0, maxLength - ellipsis.length) + ellipsis
   }